voice_assistant/requirements.txt
# Voice Assistant Platform Requirements

# Core Framework (Python 3.6 compatible versions)
fastapi==0.83.0
uvicorn==0.17.0
pydantic==1.9.2

# Configuration
PyYAML==6.0.1

# HTTP Client
httpx==0.22.0
requests==2.27.1

# Audio Processing (optional - for real implementations)
# librosa==0.10.1
# soundfile==0.12.1
# pyaudio==0.2.11

# Machine Learning (optional - for real implementations)
# torch==2.1.0
# transformers==4.35.0
# whisper==1.1.10

# Text-to-Speech (optional - for real implementations)
# gTTS==2.4.0
# pyttsx3==2.90

# Speech-to-Text (optional - for real implementations)
# vosk==0.3.45
# speechrecognition==3.10.0

# MQTT (optional)
paho-mqtt==1.6.1

# Utilities
python-multipart==0.0.5
python-dotenv==0.19.2

# Development
pytest==6.2.5
pytest-asyncio==0.15.1
black==21.12b0
flake8==4.0.1
